Combine generate files and print comments there
